Add Spring Security login page and base simple config

+++ src/main/java/ua/vspelykh/atm/model/dto/AccountDTO.java
@@ -15,7 +15,7 @@ import java.time.LocalDate;
 @SuperBuilder
 public class AccountDTO extends AbstractDTO {
     private String accountNumber;
-    private BigDecimal balance;
+    private Integer balance;
     private Integer userId;
     private String password;
     private LocalDate issueDate;
